refactor: change folder and file names
