+++ packages/fiori-mcp-server/src/tools/run-rta-workflow-step/browser/index.ts
@@ -1,0 +1,2 @@
+export { callFrontendAction, disconnectSite, stopBrowser } from './playwright-bridge';
+export type { FrontendActionResult } from './playwright-bridge';
